update refund, progress

+++ EduQuest_Domain/Entities/CourseLearner.cs
@@ -11,7 +11,7 @@ namespace EduQuest_Domain.Entities
         public bool IsActive { get; set; }
         public string? CurrentLessonId { get; set; } 
         public string? CurrentMaterialId { get; set; }
-        public int? TotalTime { get; set; }
+        public double? TotalTime { get; set; }
         public decimal? ProgressPercentage { get; set; }
 
 		[JsonIgnore]
